--- a/ModCallProcess.cpp
+++ b/ModCallProcess.cpp
@@ -6,41 +6,50 @@ ModCallProcess::ModCallProcess(ModCallParameters params){
     // parsing ref fasta
     std::time_t begin= time(NULL);
     std::cerr<< "reading reference ... ";
-    std::map<std::string, std::string> chrString;
-    MethFastaParser MethFastaParser(params.fastaFile, chrString);
+    std::vector<ReferenceChromosome> chrInfo;
+    MethFastaParser MethFastaParser(params.fastaFile, chrInfo);
     std::cerr<< difftime(time(NULL), begin) << "s\n";
+    
+    // record all ModCall result
+    std::map<std::string,std::ostringstream> chrModCallResult;
+    // Initialize an empty map in chrModCallResult to store all ModCall results.
+    // This is done to prevent issues with multi-threading, by defining an empty map first.
+    for (auto chrIter = chrInfo.begin(); chrIter != chrInfo.end(); chrIter++) {
+        chrModCallResult[chrIter->name] = std::ostringstream();
+    }
 
-    bool isFirstchr = true;
+    // set chrNumThreads and bamParserNumThreads based on parameters
+    int chrNumThreads,bamParserNumThreads;
+    setModcallNumThreads(params.numThreads, chrNumThreads, bamParserNumThreads);
+    // setNumThreads(chrInfo.size(), params.numThreads, chrNumThreads, bamParserNumThreads);
+    begin = time(NULL);
+
     //loop all chromosomes
-    for(auto chrIter = chrString.begin(); chrIter != chrString.end(); chrIter++){
-        
+    #pragma omp parallel for schedule(dynamic) num_threads(chrNumThreads)
+    for(auto chrIter = chrInfo.begin(); chrIter != chrInfo.end(); ++chrIter) {
+        std::time_t chrbegin = time(NULL);
+
         // store variant
         std::vector<ReadVariant> fReadVariantVec;
         std::vector<ReadVariant> rReadVariantVec;
         std::vector<ReadVariant> readVariantVec;
         // record hetero metyhl position
         std::map<int,int> *passPosition = new std::map<int,int>;
+
+        std::string chrName = chrIter->name;
+        std::string chrSeq = chrIter->sequence;
+        int chrLen = chrIter->length;
         
-        begin = time(NULL);
-        std::string chrName = (*chrIter).first;
-        int chrLen = (*chrIter).second.length();
+        MethBamParser *methbamparser = new MethBamParser(params, chrSeq);
         
-        std::cerr<<"parsing contig/chromosome: "<<chrName<<" ... ";
-        MethBamParser *methbamparser = new MethBamParser(params, (*chrIter).second);
-        
-        std::cerr<<"parsing Bam " << " ... ";
-        methbamparser->detectMeth(chrName, chrLen, readVariantVec);
+        methbamparser->detectMeth(chrName, chrLen, bamParserNumThreads, readVariantVec);
 
         //new new calculate depth
-        std::cerr<<"cal depth " << " ... ";
         methbamparser->calculateDepth();
         
         //judge methylation genotype
-        std::cerr<<"judge meth " << " ... ";
         methbamparser->judgeMethGenotype(chrName, readVariantVec, fReadVariantVec, rReadVariantVec);
-        
-        std::cerr<< "run algorithm ... ";
-        
+
         MethylationGraph *fGraph = new MethylationGraph(params);
         MethylationGraph *rGraph = new MethylationGraph(params);
         fGraph->addEdge(fReadVariantVec);
@@ -53,13 +62,9 @@ ModCallProcess::ModCallProcess(ModCallParameters params){
         rGraph->destroy();
         delete fGraph;
         delete rGraph;
-
         
-        //write to vcf file
-        std::cerr<<"write vcf " << " ... ";
-        methbamparser->writeResultVCF(chrName, chrString, isFirstchr, (*passPosition));
-        isFirstchr = false;
-        std::cerr<< difftime(time(NULL), begin) << "s\n";
+        //push result to ModCallResult
+        methbamparser->exportResult(chrName, chrSeq, chrLen, (*passPosition), chrModCallResult[chrName]);
         
         passPosition->clear();
         
@@ -72,7 +77,16 @@ ModCallProcess::ModCallProcess(ModCallParameters params){
         fReadVariantVec.shrink_to_fit();
         rReadVariantVec.clear();
         rReadVariantVec.shrink_to_fit();
+
+        std::cerr<< "(" << chrName << "," << difftime(time(NULL), chrbegin) << "s)";
     }
+    std::cerr<< "\nmodcall total:  " << difftime(time(NULL), begin) << "s\n";
+
+    begin= time(NULL);
+    std::cerr<<"write vcf " << " ... ";
+    //write to vcf file
+    writeResultVCF(params, chrInfo, chrModCallResult);
+    std::cerr<< difftime(time(NULL), begin) << "s\n";
 }
 
 ModCallProcess::~ModCallProcess(){